make hub page available only for XOA users
